% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaegl51.R
\docType{data}
\name{aaegl51_description}
\alias{aaegl51_description}
\title{VectorBase Gene Description}
\format{a \code{tibble} with 3 column which are:
\describe{
  \item{gene}{VectorBase gene symbol}
  \item{description}{Gene product name or description}
  \item{gene_name}{Gene name}
}}
\source{
\url{http://biomart.vectorbase.org/biomart/}
}
\usage{
aaegl51_description
}
\description{
This dataset contains the description and names for
  Aedes aegypti genes.
}
\keyword{datasets}
