% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aaegl51_symbol_conversion}
\alias{aaegl51_symbol_conversion}
\title{aaegL5.1 Gene symbol Conversion Dataset}
\format{a \code{tibble} with X column which are:
\describe{
\item{gene}{gene symbol}
\item{vectorbase}{gene symbol used by vectorbase}
\item{ncbi}{gene symbol used by ncbi}
}}
\usage{
aaegl51_symbol_conversion
}
\description{
This dataset is used to convert gene symbols used in different
             databases, e.g. NCBI Gene, KEGG, emsembl and VectorBase.
}
\keyword{datasets}
