% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_conversion.R
\docType{data}
\name{annotation_conversion}
\alias{annotation_conversion}
\title{Gene Annotation Coversion Table}
\format{a \code{tibble} with 2 column which are:
\describe{
  \item{gene_aaegl35}{Gene stable identifiers for the AaegL3.5 gene models}
  \item{gene_aaegl51}{Gene stable identifiers for the AaegL5.1 gene models}
}}
\usage{
annotation_conversion
}
\description{
This dataset contains the conversion of stable gene
             identifiers for different versions of
             Aedes aegypti genome.
}
\keyword{datasets}
