% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_gene_sets.R
\docType{data}
\name{go_gene_sets}
\alias{go_gene_sets}
\title{Gene Ontology Gene Sets}
\format{a \code{tibble} with 3 column which are:
\describe{
  \item{gene_ontology}{Gene Ontology with code}
  \item{gene}{VectorBase gene symbol}
  \item{go_domain}{Gene Ontology domain}
}}
\source{
\url{http://biomart.vectorbase.org/biomart/martview?VIRTUALSCHEMANAME=vb_mart_1806&ATTRIBUTES=alvpagwg_eg_gene.default.feature_page.ensembl_gene_id|alvpagwg_eg_gene.default.feature_page.go_id|alvpagwg_eg_gene.default.feature_page.name_1006|alvpagwg_eg_gene.default.feature_page.namespace_1003&FILTERS=&VISIBLEPANEL=resultspanel}
}
\usage{
go_gene_sets
}
\description{
This dataset contains the Aedes aegypti genes from
  each Gene Ontology in a tidy format. AaegL5.1
}
\keyword{datasets}
