% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_gene_sets.R
\docType{data}
\name{go_gene_sets_list}
\alias{go_gene_sets_list}
\title{Gene Ontology Gene Sets}
\format{a named \code{list} which each element is a Gene Ontology,
  containg a vector of the VectorBase genes for each GO}
\source{
\url{http://biomart.vectorbase.org/biomart/martview?VIRTUALSCHEMANAME=vb_mart_1806&ATTRIBUTES=alvpagwg_eg_gene.default.feature_page.ensembl_gene_id|alvpagwg_eg_gene.default.feature_page.go_id|alvpagwg_eg_gene.default.feature_page.name_1006|alvpagwg_eg_gene.default.feature_page.namespace_1003&FILTERS=&VISIBLEPANEL=resultspanel}
}
\usage{
go_gene_sets_list
}
\description{
This dataset contains the Aedes aegypti genes from
  each Gene Ontology in a list format,
  required for some bioconductor packages.
}
\keyword{datasets}
