% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg_gene_sets.R
\docType{data}
\name{kegg_gene_sets}
\alias{kegg_gene_sets}
\title{KEGG Pathways Gene Sets}
\format{a \code{tibble} with 2 column which are:
\describe{
  \item{pathway}{KEGG Pathway name with code}
  \item{gene}{VectorBase gene symbol}
}}
\source{
\url{https://www.kegg.jp/kegg-bin/download_htext?htext=aag00001&format=json&filedir=}
}
\usage{
kegg_gene_sets
}
\description{
This dataset contains the Aedes aegypti genes from
  each KEGG Pathway in a tidy format.AaegL5.1
}
\keyword{datasets}
