% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg_gene_sets.R
\docType{data}
\name{kegg_gene_sets_list}
\alias{kegg_gene_sets_list}
\title{KEGG Pathways Gene Sets}
\format{a named \code{list} which each element is a KEGG Pathway, containg a
  vector of the VectorBase genes for each pathway}
\source{
\url{https://www.kegg.jp/kegg-bin/download_htext?htext=aag00001&format=json&filedir=}
}
\usage{
kegg_gene_sets_list
}
\description{
This dataset contains the Aedes aegypti genes from
  each KEGG Pathway in a list format,
  required for some bioconductor packages. AaegL5.1
}
\keyword{datasets}
