% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaegl51.R
\docType{data}
\name{transcript_to_gene}
\alias{transcript_to_gene}
\title{Transcript to Gene Level Conversion Table}
\format{a \code{tibble} with 2 column which are:
\describe{
  \item{TXNAME}{VectorBase transcript}
  \item{GENEID}{VectorBase gene}
}}
\source{
\url{http://biomart.vectorbase.org/biomart/}
}
\usage{
transcript_to_gene
}
\description{
Dataset used to convert transcript level to gene level analysis
  for the Aedes aegypti AaegL5.1 - TX2GENE.
}
\keyword{datasets}
