% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaegdata-package.R
\docType{package}
\name{aaegdata-package}
\alias{aaegdata}
\alias{aaegdata-package}
\title{aaegdata: Aedes aegypti Mosquitoes Molecular Information}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

This is a data package containing Aedes aegypti vector mosquitoes 
             biologically relevant molecular and biochemical data. Data colected from 
             public available databases and presented here in a clean, structured and
             ready to use format.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/luciorq/aaegdata}
  \item Report bugs at \url{https://github.com/luciorq/aaegdata/issues}
}

}
\author{
\strong{Maintainer}: Lucio Rezende Queiroz \email{luciorqueiroz@gmail.com} (0000-0002-6090-1834)

}
\keyword{internal}
